Module D.Array1

FFT of Bigarray.Array1.

val create : ('a, 'b) Stdlib.Bigarray.kind -> 'c Stdlib.Bigarray.layout -> int -> ('a, 'b, 'c) Stdlib.Bigarray.Array1.t

See Fftw3.Sig.Genarray.create.

val of_array : ('a, 'b) Stdlib.Bigarray.kind -> 'c Stdlib.Bigarray.layout -> 'a array -> ('a, 'b, 'c) Stdlib.Bigarray.Array1.t

of_array kind layout a build a one-dimensional aligned big array initialized from the given array.

type 'l complex_array = (Stdlib.Complex.t, complex_elt, 'l) Stdlib.Bigarray.Array1.t

Double precision complex 1D array.

type 'l float_array = (float, float_elt, 'l) Stdlib.Bigarray.Array1.t

Double precision float 1D array.

val dft : dir -> ?⁠meas:measure -> ?⁠destroy_input:bool -> ?⁠unaligned:bool -> ?⁠howmany_n:int array -> ?⁠howmanyi:int list -> ?⁠ni:int -> ?⁠ofsi:int -> ?⁠inci:int -> 'l complex_array -> ?⁠howmanyo:int list -> ?⁠no:int -> ?⁠ofso:int -> ?⁠inco:int -> 'l complex_array -> c2c plan

dft dir x y returns a plan to compute the DFT of x and store it in y.

The parameters meas, destroy_input, unaligned are as for Fftw3.Sig.Genarray.dft.

parameter n

the logical length of the array. If not provided, it is automatically computed from ofsi, inci and Array1.dim x.

Remark: If you want to transform several 1D arrays at once, use Fftw3.Sig.Array2.dft with ~howmanyi and ~howmanyo set appropriately.

val r2c : ?⁠meas:measure -> ?⁠destroy_input:bool -> ?⁠unaligned:bool -> ?⁠howmany_n:int array -> ?⁠howmanyi:int list -> ?⁠ni:int -> ?⁠ofsi:int -> ?⁠inci:int -> 'l float_array -> ?⁠howmanyo:int list -> ?⁠no:int -> ?⁠ofso:int -> ?⁠inco:int -> 'l complex_array -> r2c plan

See Fftw3.Sig.Genarray.r2c.

val c2r : ?⁠meas:measure -> ?⁠destroy_input:bool -> ?⁠unaligned:bool -> ?⁠howmany_n:int array -> ?⁠howmanyi:int list -> ?⁠ni:int -> ?⁠ofsi:int -> ?⁠inci:int -> 'l complex_array -> ?⁠howmanyo:int list -> ?⁠no:int -> ?⁠ofso:int -> ?⁠inco:int -> 'l float_array -> c2r plan

See Fftw3.Sig.Genarray.c2r.

val r2r : r2r_kind -> ?⁠meas:measure -> ?⁠destroy_input:bool -> ?⁠unaligned:bool -> ?⁠howmany_n:int array -> ?⁠howmanyi:int list -> ?⁠ni:int -> ?⁠ofsi:int -> ?⁠inci:int -> 'l float_array -> ?⁠howmanyo:int list -> ?⁠no:int -> ?⁠ofso:int -> ?⁠inco:int -> 'l float_array -> r2r plan

See Fftw3.Sig.Genarray.r2r.