Module S.Guru

Guru execution of plans.

If you want to transform other arrays than those specified in the plan, you are advised to create a new plan — it won't be too expensive if the wisdom can be reused. To transform a known bunch of arrays of the same size, you should not use the following functions but instead create a plan with ?howmany set appropriately.

These functions are thread safe. You can even execute the same plan in parallel by multiple threads by providing different arrays than the ones with which the plan was created.