Module Fftw3.Wisdom

Managing wisdom. Save and restore plans to/from disk or other media.

val export : (char -> unit) -> unit

Wisdom.export write exports the current wisdom to any medium, as specified by the callback function write.

This function is not thread safe.

val to_file : string -> unit

Wisdom.to_file fname writes the current wisdom to the file fname.

val to_string : unit -> string

Wisdom.to_string() exports the current wisdom as a string.

val import : (unit -> char) -> unit

Wisdom.import read imports wisdom from any input medium, as specified by the callback function read. If the end of the input data is reached (which should never happen for valid data), read should raise End_of_file. The imported wisdom replaces any wisdom accumulated by the running program.

This function is not thread safe.

raises Failure

if the wisdom was not successfully read.

val from_file : string -> unit

Widsom.from_file fname replace the current wisdom with the one read from the file fname.

raises Failure

if the wisdom was not successfully read.

val from_string : string -> unit

Wisdom.from_string s replace the current wisdom whith the one read from s.

raises Failure

if the wisdom was not successfully read.

val from_system : unit -> unit

Wisdom.from_system() replace the current wisdom with one read from an implementation-defined standard file (e.g. /etc/fftw/wisdom).

raises Failure

if the wisdom was not successfully read.

val forget : unit -> unit

Wisdom.forget() causes all accumulated wisdom to be discarded. (New wisdom can be gathered subsequently though.)