Module Fftw3__Fftw3_geomF

Geometry checks. Uniform treatement of the C and FORTRAN layouts through macros (for the bigarray interface). Does not depend on the precision of the arrays.

val get_geom : string -> string -> int array option -> string -> int array option -> string -> int array option -> ('a, 'b, 'c) Stdlib.Bigarray.Genarray.t -> int * int array * int array * int array * int array

val only_ones : int array -> bool

only_ones d tells whether d entries are all 1 -- which is in particular the case if d is empty.

val get_geom_hm : string -> string -> int array -> string -> int array Fftw3__.Fftw3_utils.List.t -> int array -> int array -> ('a, 'b, 'c) Stdlib.Bigarray.Genarray.t -> int array * int array

Check whether the matrix given by hm_n (howmany matrix) and hm is a valid submatrix of the hermitian matrix mat.

returns

the hm_stride, hm_n (howmany matrix), stride and n (logical dimensions).

val apply : string -> (int -> int -> 'a -> int array -> int array -> int array -> int array -> int array -> 'b) -> int array -> int array Fftw3__.Fftw3_utils.List.t -> ?⁠ni:int array -> int array option -> int array option -> ('c, 'd, 'e) Stdlib.Bigarray.Genarray.t -> int array Fftw3__.Fftw3_utils.List.t -> ?⁠no:int array -> int array option -> int array option -> ('f, 'g, 'h) Stdlib.Bigarray.Genarray.t -> logical_dims:(int array -> int array -> string -> 'a) -> 'b